refactor(storage): separate public types from migration internals
Audit roadmap section 4.1.4 (storage type hygiene). Types/exports only;
zero runtime behavior change.

- Add lib/storage/public-types.ts holding the current-version shapes that
  external consumers and lib/index.ts need: CooldownReason,
  RateLimitStateV3, AccountMetadataV3, AccountStorageV3 (moved verbatim
  from lib/storage/migrations.ts). The lib/storage.ts facade re-exports
  them as before, so the published ./storage subpath surface for these
  names is unchanged.
- lib/storage/migrations.ts now contains only the historical v1 shapes
  (AccountMetadataV1, AccountStorageV1) plus migrateV1ToV3, importing the
  current shapes from public-types.ts.
- Drop AccountMetadataV1 and AccountStorageV1 from the facade's export
  list: nothing outside lib/storage/migrations.ts (and the facade's own
  internal migrateV1ToV3 call) imports them, and
  test/public-api-contract.test.ts pins only runtime exports and package
  subpaths, not these types.
- Update the direct deep importers of AccountMetadataV3 to the new home:
  lib/codex-manager/account-pool-write.ts and
  test/codex-manager-account-pool-write.test.ts.
- Note: the audit proposed also removing RateLimitStateV3 from the
  facade, but that premise is stale; it is part of the current v3 shape
  (AccountMetadataV3.rateLimitResetTimes) and is consumed via the facade
  by lib/accounts.ts, so it stays public in public-types.ts.
